feat: Implement task lease and retry mechanism in task management
- Added new columns to the tasks table for lease management: lease_until, lease_renewed_at, and retry_count.
- Introduced SQL index for efficient querying of tasks based on status and lease information.
- Enhanced AppConfig to include parameters for task lease duration and retry limits.
- Updated the DistributedService to handle task heartbeats for lease renewal.
- Implemented new API endpoints for requeuing and rerunning tasks, improving task management capabilities.

+++ backend/src/services/cache.rs
@@ -160,7 +160,7 @@ impl FileHashCache {
         let mut sys = System::new_all();
         sys.refresh_all();
 
-        let available_memory = sys.available_memory() as u64;
+        let available_memory = sys.available_memory();
 
         // 为哈希缓存分配可用内存的5%，每个缓存项大约200字节
         let cache_memory_mb = (available_memory / 1024 / 1024) as usize / 20; // 5%
